Create Project with Preferred Method
This Project differs from the main project folder because this implements a method called "chooseByPreffered". This takes the absent teacher's preffered substitute and puts it in the time slot. The other Project folder chooses substitutes by looking at their teachables. The reason for the separation of files is to demonstrate both methods working correctly and independently.