Add files via upload
This version of the project differs from the main Project folder. This versions driver uses a method called chooseByPreffered which for each absence, chooses the absent teacher's preferred teacher as the teacher for the timeslot. If this teacher is unavailable, we set the program to choose a random, available teacher. The main Project differs because it implements a method that chooses by teachables instead. We thought it was best to split into to different projects so that the functionality of both systems could be shown independently.